also ignore firmware files on scan that match
