+++ backend/handler/filesystem/base_handler.py
@@ -1,5 +1,6 @@
 import os
 import re
+import fnmatch
 from abc import ABC
 from enum import Enum
 from typing import Final
@@ -113,3 +114,28 @@ class FSHandler(ABC):
     def parse_file_extension(self, file_name) -> str:
         match = re.search(EXTENSION_REGEX, file_name)
         return match.group(1) if match else ""
+    
+    def _exclude_files(self, files, filetype) -> list[str]:
+        cnfg = cm.get_config()
+        excluded_extensions = getattr(cnfg, f"EXCLUDED_{filetype.upper()}_EXT")
+        excluded_names = getattr(cnfg, f"EXCLUDED_{filetype.upper()}_FILES")
+        excluded_files: list = []
+
+        for file_name in files:
+            # Split the file name to get the extension.
+            ext = self.parse_file_extension(file_name)
+
+            # Exclude the file if it has no extension or the extension is in the excluded list.
+            if not ext or ext in excluded_extensions:
+                excluded_files.append(file_name)
+
+            # Additionally, check if the file name mathes a pattern in the excluded list.
+            if len(excluded_names) > 0:
+                [
+                    excluded_files.append(file_name)
+                    for name in excluded_names
+                    if file_name == name or fnmatch.fnmatch(file_name, name)
+                ]
+
+        # Return files that are not in the filtered list.
+        return [f for f in files if f not in excluded_files]
